Crate claw_parse

claw_parse

This crate is part of the Claw programming language. You can see the project README here.

Re-exports

pub use crate::ast::AST;

pub use crate::parser::create_parser;

Modules

ast

Contains the abstract syntax tree definitions.

parser

Contains the logic of the actual parser, written using [chumsky].

Functions

build_report

Transforms a result to use [ariadne]’s [Report].

get_source

Returns the [Source] of a string reference.

parse

Parses a str into AST.

Type Definitions

ParseError

Error type returned by the parsers.